Style fixes.
